Astro smart bundling

+++ runtimes/node/versions/latest/helpers/astro/bundle.sh
@@ -4,5 +4,8 @@ if [ -n "$OPEN_RUNTIMES_OUTPUT_DIRECTORY" ]; then
     cd $OPEN_RUNTIMES_OUTPUT_DIRECTORY
 fi
 
-mv /usr/local/build/package*.json ./
-mv /usr/local/build/node_modules/ ./node_modules/
+ENTRYPOINT="./server/entry.mjs"
+if [ -e "$ENTRYPOINT" ]; then
+    mv /usr/local/build/package*.json ./
+    mv /usr/local/build/node_modules/ ./node_modules/
+fi
